+++ src/Tools/OBJ2OMPF/src/OBJ2OMPF.cpp
@@ -82,7 +82,7 @@ void OBJ2OMPF::convert(int argc, char* argv[])
 {
     if (analyzeParam(argc, argv))
     {
-        int32 materialID = iMaterialResourceFactory::getInstance().createMaterial("Textured");
+        uint64 materialID = iMaterialResourceFactory::getInstance().createMaterial("Textured");
         iMaterialResourceFactory::getInstance().getMaterial(materialID)->addShaderSource(L"textured.vert", iShaderObjectType::Vertex);
         iMaterialResourceFactory::getInstance().getMaterial(materialID)->addShaderSource(L"textured_directional_light.frag", iShaderObjectType::Fragment);
         iMaterialResourceFactory::getInstance().getMaterial(materialID)->getRenderStateSet().setRenderState(iRenderState::Texture2D0, iRenderStateValue::On);
